move lib/api -> lib/
