+++ lib/channel.js
