+++ lib/entry.js
