+++ lib/initialize.js
