+++ lib/signal.js
