+++ lib/space.js
