+++ lib/window.js
