+++ test/unit/window.spec.js
@@ -1,4 +1,4 @@
-const createWindow = require('../../lib/api/window')
+const createWindow = require('../../lib/window')
 
 const browserWindow = window
 const { Event } = browserWindow
